requirements.txt
discord.py
asyncio
requests
psycopg2-binary
PyNaCl
youtube-dl
cachetools
janus
